Moved most parts of the GamePage to newly created UserControls to improve readability and structure.
