+++ src/Codebreaker.WinUI/Views/Components/GamePage/Moves.xaml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<UserControl
+    x:Class="CodeBreaker.WinUI.Views.Components.GamePage.Moves"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:CodeBreaker.WinUI.Views.Components.GamePage"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <ScrollViewer Padding="0,0,0,15" x:Name="ScrollViewer">
+        <ListBox 
+        x:Name="listGameMoves"
+        Background="Transparent"
+        SelectedIndex="-1"
+        IsHitTestVisible="False"
+        ItemsSource="{x:Bind ViewModel.Game.Moves, Mode=OneWay}"
+        ItemTemplate="{StaticResource PegsTemplate}" />
+    </ScrollViewer>
+</UserControl>
