+++ src/Codebreaker.WinUI/Views/Components/GamePage/StartGameComponent.xaml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="utf-8"?>
+<UserControl
+    x:Class="CodeBreaker.WinUI.Views.Components.GamePage.StartGameComponent"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:CodeBreaker.WinUI.Views.Components.GamePage"
+    xmlns:toolkitConverters="using:CommunityToolkit.WinUI.UI.Converters"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:cme="using:CodeBreaker.WinUI.CustomMarkupExtensions"
+    mc:Ignorable="d">
+    <UserControl.Resources>
+        <toolkitConverters:BoolToVisibilityConverter x:Key="BoolToVisibility" />
+    </UserControl.Resources>
+
+    <Grid ColumnSpacing="25">
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition />
+            <ColumnDefinition Width="Auto" />
+        </Grid.ColumnDefinitions>
+        <Grid Grid.Column="0">
+            <!-- The extra grid and the empy (enabled) TextBlock are necessary for showing the tooltip over the (disabled) TextBox. -->
+            <TextBlock />
+            <TextBox
+            Header="{cme:ResourceString Name=GamePage_NameInputHeader}"
+            PlaceholderText="{cme:ResourceString Name=GamePage_NameInputPlaceholder}"
+            VerticalAlignment="Center"
+            Text="{x:Bind ViewModel.Username, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
+        </Grid>
+        <Button
+        Grid.Column="1"
+        VerticalAlignment="Bottom"
+        HorizontalAlignment="Stretch"
+        Padding="35,8"
+        Command="{x:Bind ViewModel.StartGameCommand, Mode=OneTime}">
+            <StackPanel Orientation="Vertical" Spacing="5">
+                <SymbolIcon Symbol="Play" Visibility="{x:Bind ViewModel.StartGameCommand.IsRunning, Mode=OneWay, Converter={StaticResource BoolToVisibility}, ConverterParameter=True}" />
+                <ProgressRing Visibility="{x:Bind ViewModel.StartGameCommand.IsRunning, Mode=OneWay}" Width="20" Height="20" />
+                <TextBlock Text="{cme:ResourceString Name=GamePage_StartButtonText}" />
+            </StackPanel>
+        </Button>
+    </Grid>
+</UserControl>
