+++ src/Codebreaker.WinUI/Views/Pages/GamePage.xaml.cs
@@ -1,8 +1,6 @@
 using Codebreaker.ViewModels;
 using Codebreaker.ViewModels.Messages;
 using CommunityToolkit.Mvvm.Messaging;
-using Microsoft.UI.Xaml.Media.Animation;
-using Microsoft.UI.Xaml.Shapes;
 
 namespace CodeBreaker.WinUI.Views.Pages;
 
@@ -10,7 +8,6 @@ namespace CodeBreaker.WinUI.Views.Pages;
 /// An empty page that can be used on its own or navigated to within a Frame.
 /// </summary>
 public sealed partial class GamePage : Page,
-    IRecipient<MakeMoveMessage>,
     IRecipient<GameStartedMessage>,
     IRecipient<GameEndedMessage>,
     IRecipient<GameCancelledMessage>
@@ -20,8 +17,7 @@ public sealed partial class GamePage : Page,
     public GamePage()
     {
         ViewModel = App.GetService<GamePageViewModel>();
-        WeakReferenceMessenger.Default.RegisterAll(this);
-        WeakReferenceMessenger.Default.UnregisterAllOnUnloaded(this);
+        WeakReferenceMessenger.Default.RegisterAllAndUnregisterAllOnUnloaded(this);
         InitializeComponent();
         this.GoToState("Start", false);
     }
@@ -33,44 +29,8 @@ public sealed partial class GamePage : Page,
     {
         this.GoToState("Finished");
         await Task.Delay(1000);
-        ScrollPegListToBottom();
     }
 
-    public void Receive(MakeMoveMessage message)
-    {
-        // Move must be completed
-        if (!message.IsSet)
-            return;
-
-        var animationService = ConnectedAnimationService.GetForCurrentView();
-        animationService.DefaultDuration = TimeSpan.FromMilliseconds(500);
-        listGameMoves
-            .ItemContainerGenerator
-            .ContainerFromIndex(listGameMoves.Items.Count - 1)
-            .FindChildrenRecursively<Ellipse>()
-            .Foreach((ellipse, i) =>
-            {
-                ConnectedAnimation? animation = animationService.GetAnimation($"guess{i}");
-
-                // No animation found for this ellipxe -> the ellipse is most likely a key-peg
-                if (animation is null)
-                    return;
-
-                animation.Configuration = new BasicConnectedAnimationConfiguration();
-                animation.TryStart(ellipse);
-            });
-
-        ScrollPegListToBottom();
-    }
-
-    private void ScrollPegListToBottom()
-    {
-        PegScrollViewer.UpdateLayout();
-        PegScrollViewer.ChangeView(null, PegScrollViewer.ScrollableHeight, null, false);
-    }
-
-    public void Receive(GameCancelledMessage message)
-    {
+    public void Receive(GameCancelledMessage message) =>
         this.GoToState("Start");
-    }
 }
